Modificada a forma de chamar os métodos que calculam se um evento ocorreu ou não, agora usa algo semelhante ao padrão Observer

+++ vrpn-ufrn-team-client/ConfigFileReader.c
@@ -77,7 +77,7 @@ bool ConfigFileReader::readConfigFile(char * fileName,
 		}
 	}
 
-	for ( int i = 0; i < devs.size(); i++ ) {
+	for ( size_t i = 0; i < devs.size(); i++ ) {
 		
 		if ( setDevType(js,"kinect",DEVTYPE_KINECT,devs.at(i)) ) {
 			continue;
